requirements.txt
alabaster==0.7.16
babel==2.16.0
blinker==1.8.2
chemprop==1.6.1
click==8.1.7
cloudpickle==3.0.0
docstring_parser==0.16
docutils==0.21.2
et-xmlfile==1.1.0
Flask==3.0.3
fsspec==2024.9.0
future==1.0.0
hyperopt==0.2.7
imagesize==1.4.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
joblib==1.4.2
matplotlib==3.9.2
mkl-service==2.4.0
mypy-extensions==1.0.0
openpyxl==3.1.5
pandas-flavor==0.6.0
patsy==1.0.1
protobuf==5.28.2
py4j==0.10.9.7
Pygments==2.18.0
pyparsing==3.1.4
rdkit==2024.3.5
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.13.1
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
statsmodels==0.14.4
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tomli==2.0.2
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typed-argument-parser==1.10.1
typing-inspect==0.9.0
Werkzeug==3.0.4
xarray==2024.7.0
